Fix for platform where sizeof(int) != sizeof(long)
When building the current code on a platform where int and long are not
the same size, spy_message do not have the correct structure definition.

For example building on Ubuntu 16.04 x86_64 result in some spy_message
not being corrected formatted.

The ArbIDOrHeader also changed from signed to unsigned to match the
included icsnVC40 header file.