Handle member types prefixed with struct keyword
In C, if you define a struct as `typedef struct {} abc_t;`, you can
then use it as just `abc_t foo;`.

If, however, you define it as `struct abc_t {};`, it must be used
as `struct abc_t foo;`.

Some macOS structures were defined in the second way.

When we were just looking for the struct keyword to denote
a new subobject instead of a member variable, this would
cause parsing of the structures to fail.

Ultimately, no structures would be generated, which caused
a "_getPythonModuleObject(): Failed to import module"
error when trying to call get_device_settings.

Address issue #107 of the python_ics github project.